refactor: split oversized modules for <500 LOC per file
Split three files that exceeded 500 LOC:

- rlm_runtime_modules.py (695→95 shim) → builders.py (500) + registry.py (238)
  Separates DSPy module classes and RLM factory from the registry/lookup layer.
  Per DSPy docs: dspy.Module subclasses and dspy.RLM factory in builders,
  RuntimeModuleDefinition registry pattern in registry.

- sandbox_delegate_tools.py (655→458) → batch_tools.py (241)
  Extracts parallel_semantic_map batch orchestration. Per DSPy docs:
  dspy.Tool wrappers kept with their build_*_tools() builders.

- sandbox_storage_tools.py (621→374) → buffer_tools.py (189) + process_tools.py (139)
  Separates buffer/volume I/O from process management.

All original import paths preserved via backwards-compat re-exports.
